requirements-inference.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

pillow==10.4.0
opencv-python-headless==4.9.0.80

numpy==1.26.4
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.13.1
matplotlib==3.8.4
seaborn==0.13.2

torch==2.3.1
torchvision==0.18.1
pytorch-lightning==2.3.3
grad-cam==1.4.8

mlflow==2.12.1
boto3==1.42.71
pyarrow==15.0.2

pydantic==2.9.0
rich==14.2.0
pyyaml==6.0.3
zstandard==0.24.0

python-magic==0.4.27
kornia==0.7.2
PyWavelets==1.8.0